@@ -1,15 +1,14 @@
 // Copyright (c) 2017 PlanGrid, Inc.
 
-import React, { Component } from 'react';
+import React, { Component } from 'react'
 
-import ReactDataGrid from 'react-data-grid';
-import CSV from 'comma-separated-values';
+import ReactDataGrid from 'react-data-grid'
+import CSV from 'comma-separated-values'
 
 class CsvViewer extends Component {
-
   static parse(data) {
-    const rows = [];
-    const columns = [];
+    const rows = []
+    const columns = []
 
     new CSV(data).forEach((array) => {
       if (columns.length < 1) {
@@ -20,40 +19,41 @@ class CsvViewer extends Component {
             resizable: true,
             sortable: true,
             filterable: true,
-          });
-        });
+          })
+        })
       } else {
-        const row = {};
+        const row = {}
         array.forEach((cell, idx) => {
-          row[`key-${idx}`] = cell;
-        });
-        rows.push(row);
+          row[`key-${idx}`] = cell
+        })
+        rows.push(row)
       }
-    });
+    })
 
-    return { rows, columns };
+    return { rows, columns }
   }
 
   constructor(props) {
-    super(props);
-    this.state = CsvViewer.parse(props.data);
+    super(props)
+    this.state = CsvViewer.parse(props.data)
   }
 
-  componentWillReceiveProps(nextProps) {
-    this.setState(CsvViewer.parse(nextProps.data));
+  UNSAFE_componentWillReceiveProps(nextProps) {
+    // TODO
+    this.setState(CsvViewer.parse(nextProps.data))
   }
 
   render() {
-    const { rows, columns } = this.state;
+    const { rows, columns } = this.state
     return (
       <ReactDataGrid
         columns={columns}
         rowsCount={rows.length}
-        rowGetter={i => rows[i]}
+        rowGetter={(i) => rows[i]}
         minHeight={this.props.height || 650}
       />
-    );
+    )
   }
 }
 
-export default CsvViewer;
+export default CsvViewer
