@@ -1,31 +1,32 @@
-// Copyright (c) 2017 PlanGrid, Inc.
+// TODO
+/* eslint-disable jsx-a11y/media-has-caption */
 
-import React, { Component } from 'react';
+import React, { Component } from 'react'
 
-import 'styles/video.scss';
-import Loading from '../loading';
+import 'styles/video.scss'
+import Loading from '../loading'
 
 class VideoViewer extends Component {
   constructor(props) {
-    super(props);
+    super(props)
     this.state = {
       loading: true,
-    };
+    }
   }
 
   onCanPlay() {
-    this.setState({ loading: false });
+    this.setState({ loading: false })
   }
 
   renderLoading() {
     if (this.state.loading) {
-      return <Loading />;
+      return <Loading />
     }
-    return null;
+    return null
   }
 
   render() {
-    const visibility = this.state.loading ? 'hidden' : 'visible';
+    const visibility = this.state.loading ? 'hidden' : 'visible'
     return (
       <div className="pg-driver-view">
         <div className="video-container">
@@ -34,15 +35,14 @@ class VideoViewer extends Component {
             style={{ visibility }}
             controls
             type={`video/${this.props.fileType}`}
-            onCanPlay={e => this.onCanPlay(e)}
-            src={this.props.filePath}
-          >
+            onCanPlay={(e) => this.onCanPlay(e)}
+            src={this.props.filePath}>
             Video playback is not supported by your browser.
           </video>
         </div>
       </div>
-    );
+    )
   }
 }
 
-export default VideoViewer;
+export default VideoViewer
